+++ Geoscape/ScrapVehicle/mod_info.js
@@ -12,6 +12,8 @@
       "Nexus" : "https://nexusmods.com/phoenixpoint/mods/26/",
       "GitHub" : "https://github.com/Sheep-y/PhoenixPt-Mods/",
    },
+   Requires : 
+      { Id: "Phoenix Point", Min: "1.5.2.0" },
    Disables :
       { Id: "PhoenixPt_ScrapVehicles", Max: "1.0.0.0" },
 })