+++ Geoscape/ScrapVehicle/changelog.md
@@ -6,13 +6,13 @@ Changelog of Scrap Vehicle, a Phoenix Point Mod by Sheepy
 * Fix: Vehicles now have correct name in stats panel.
 * New: Vehicles now show their location in stats panel.
 
-# Version 2.0, 2020-03-18
+# Version 2, 2020-03-18
 
 * Update to Phoenix Point 1.0.56200 and Modnix 0.90
 * Fix: Scrapping vehicle will prevent game from being saved or enter combat.
 * Fix: Last aircraft may not be scrapped, to prevent freezing the game.
 
-# Version 1.0, 2019-12-24
+# Version 1, 2019-12-24
 
 * Allow scrapping vehicles and mutogs when they are in a base.
 * Vehicles can only be scrapped if the base has a functional vehicle bay.